interview-service/requirements.txt
fastapi
uvicorn
sqlalchemy 
psycopg2-binary 
langchain 
langchain-ollama 
langchain-community 
pydantic 
python-dotenv 
